Revert improvements made after Friday deadline
NOTE:
I took a peek over the weekend and fixed a couple of flaws in the
version of the challenge that I submitted. Since Marcio & Giuseppe will
look at the GitHub repo rather than the bundled repo I sent by email,
I'm reverting those changes so that the `main` branch reflects what I
sent to Lucia on Friday.

The issues I corrected were:
- I extracted the `ArtboardList` component from the document view
- Corrected the html tag used in `PageLayout.Content` (should be `<main>`
  rather than `<section>`)
- Updated the `useArtboard` hook, so it's smart enough to know that when
  the `notFound` value is true, `artboard === null`
- I use rem units throughout the code, but my typography resets were
  not made on the root (html) element --which essentially means that
  rem-based dimensions only worked by coincidence... big whoopsie :(

These changes are in the `post-challenge-improvements` branch, if you're
interested.